pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qoa4ml"]

[tool.hatch.build.targets.sdist]
include = ["src/qoa4ml"]

[project]
name = "qoa4ml"
version = "0.3.18"
description = "Quality of Analysis for Machine Learning"
readme = "README.md"

authors = [
  { name = "AaltoSEA", email = "tri.m.nguyen@aalto.fi" },
  { name = "nguu0123", email = "nguu0123@gmail.com" },
]
license = { text = "Apache-2.0" }
keywords = ["Monitoring", "Machine Learning"]
homepage = "https://rdsea.github.io/"
repository = "https://github.com/rdsea/QoA4ML"
documentation = "https://rdsea.github.io/QoA4ML/"
requires-python = ">=3.12"
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: System :: Monitoring",
  "Topic :: Utilities",
  "Operating System :: POSIX :: Linux",
]

dependencies = [
  "pika>=1.3.2",
  "psutil>=5.9.2",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
  "fastapi>=0.110.2",
  "pydantic>=2.7.4,<3",
  "tinyflux>=1.0.0",
  "uvicorn>=0.29.0",
  "docker>=7.1.0,<8",
  "numpy>=1.26.0",
  "devtools>=0.12.0",
]

optional-dependencies.kafka = ["confluent-kafka>=2.4.0"]
optional-dependencies.ml = [
  "paho-mqtt==1.6.1",
  "Pillow>=10.0.0",
  "prometheus-client==0.16.0",
  "tensorflow>=2.15.0",
  "pandas>=2.2.2",
  # "ydata-quality==0.1.0",
]

optional-dependencies.otel = [
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-instrumentation-system-metrics>=0.47b0",
]

optional-dependencies.docs = [
  "mkdocs-material[recommended,imaging]>=9.5.17 ",
  "mkdocstrings-python",
  "cairosvg",
  "mkdocs-gen-files",
  "mkdocs-literate-nav",
  "mkdocs-section-index",
]

[project.urls]
Repository = "https://github.com/rdsea/QoA4ML"
Documentation = "https://rdsea.github.io/QoA4ML/"
AaltoSEA = "https://rdsea.github.io/"


[dependency-groups]
dev = [
  "pre-commit",
  "ruff>=0.15.7",
  "codespell",
  "mypy",
  "pytest>=8.3.2",
  "pytest-sugar>=1.0.0",
  "pytest-cov>=6.0.0",
]


[tool.ruff]
exclude = ["example/bts/"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 10

[tool.ruff.lint.pydocstyle]
convention = "pep257" # Accepts: "google", "numpy", or "pep257".

[tool.ruff.lint]
exclude = ["src/qoa4ml/utils/pynvml_forked.py", "example/bts/"]
select = [
  "F",   # Pyflakes
  "E",   # Pycode styles
  "W",   # Pycode styles
  "N",   # Pep8 naming
  "I",   # Isort
  "UP",  # Pyupgrade
  "C4",  # Flake8 comprehension
  "ISC", # Flake8 implicit str concat
  "B",   # Flake8 bugbear
  "PGH", # Pygrep hook
  "RUF", # Ruff specific rules
]
ignore = [
  "E501",   # Line too long
  "RET504", # Unnecessary assignment
  "RET505", # Unnecessary else after return
  "C901",   # Too complex function
  "B904",   # raise-without-from-inside-except, enable when understand more
  "PLR",
  "ISC001",
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
exclude = ["src/qoa4ml/utils/pynvml_forked.py"]
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disable_error_code = [
  "import-untyped",
]

[tool.pytest.ini_options]
markers = [
  "requires_docker: tests that require a running Docker daemon",
  "requires_gpu: tests that require an NVIDIA GPU",
]

[tool.coverage.run]
source = ["qoa4ml"]
omit = ["src/qoa4ml/utils/pynvml_forked.py"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.",
]

